+++ source/PascalScript.pas
@@ -20,7 +20,7 @@ uses
   ProjectTask;
 
 Type
-  TPascalScript = class
+  TPascalScript = class(tObject)
   protected
     foPlugins: TPlugins;
     fImp: TPSRuntimeClassImporter;
@@ -46,6 +46,74 @@ Type
   end;
 
 implementation
+
+function GetIncludeFullFilename(aIncludeFilename: String): string;
+Var
+  lsUnitNameFilename: String;
+begin
+  Result := aIncludeFilename;
+
+  if FileExists(aIncludeFilename) then Exit;
+
+  if TNovusFileUtils.IsJustFilenameOnly(aIncludeFilename) then
+    begin
+      if FileExists(TNovusFileUtils.TrailingBackSlash(oruntime.oProject.oProjectConfigLoader.Workingdirectory) +aIncludeFilename) then
+        Result := TNovusFileUtils.TrailingBackSlash(oruntime.oProject.oProjectConfigLoader.Workingdirectory) +aIncludeFilename
+      else
+      if FileExists(TNovusFileUtils.TrailingBackSlash(oruntime.oProject.oProjectConfigLoader.SearchPath)+ aIncludeFilename) then
+        Result := TNovusFileUtils.TrailingBackSlash(oruntime.oProject.oProjectConfigLoader.SearchPath)+ aIncludeFilename
+    end;
+
+end;
+
+function CENeedFile(Sender: TPSPreProcessor; const callingfilename: tbtstring; var FileName, Output: tbtstring): Boolean;
+var
+  lsIncludeFile: string;
+  f: TFileStream;
+begin
+  lsIncludeFile := GetIncludeFullFilename(FileName);
+  try
+    f := TFileStream.Create(lsIncludeFile, fmOpenRead or fmShareDenyWrite);
+  except
+    Result := False;
+
+    Exit;
+  end;
+
+  try
+    SetLength(Output, f.Size);
+    f.Read(Output[1], Length(Output));
+  finally
+    f.Free;
+  end;
+  Result := true;
+end;
+
+function PreProcessor(aObject: tObject; aName: tbtstring; var aScript: TStringList): TPSPreProcessor;
+Var
+  fFPP: TPSPreProcessor;
+  lsScript: tbtstring;
+begin
+  Try
+    lsScript := aScript.text;
+
+    fFPP := TPSPreProcessor.Create;
+    fFPP.Clear;
+    fFPP.Id := aObject;
+
+    fFPP.OnNeedFile := CENeedFile;
+
+    fFPP.Defines.Text := aScript.text;
+    fFpp.MainFile := aScript.text;
+    fFpp.MainFileName := aName;
+    fFpp.PreProcess(aName, lsScript);
+  Finally
+    aScript.text := lsScript;
+
+    result := fFpp;
+  End;
+end;
+
 
 procedure OnException(Sender: TPSExec; ExError: TPSError;
   const ExParam: tbtstring; ExObject: TObject; ProcNo, Position: Cardinal);
@@ -81,8 +149,9 @@ end;
 function CustomOnUses(Sender: TPSPascalCompiler;
   const Name: AnsiString): Boolean;
 Var
-  lList: TStringList;
-  FTPSCompileTimeClass: TPSCompileTimeClass;
+  fUseScript: TStringList;
+  fTPSCompileTimeClass: TPSCompileTimeClass;
+  fFpp: TPSPreProcessor;
 begin
   if Name = 'SYSTEM' then
   begin
@@ -96,19 +165,28 @@ begin
         Try
           oruntime.oAPI_Output.Log('Compiling unit ... ' + GetUnitFilename(name));
 
-          lList := TStringList.Create;
-          lList.LoadFromFile(GetUnitFilename(name));
+          fUseScript := TStringList.Create;
+          fUseScript.LoadFromFile(GetUnitFilename(name));
 
-          if Sender.Compile(lList.Text) then
+          fFpp := PreProcessor(NIL, Name, fUseScript);
+
+          if Sender.Compile(fUseScript.Text) then
           begin
             Result := true;
+
           end
           else
           begin
             Result := False;
+
+//            Sender.MakeError(Sender.UnitName, ecUnknownIdentifier, Name);
           end;
         Finally
-          lList.Free;
+          fFpp.AdjustMessages(Sender);
+
+          fUseScript.Free;
+
+          fFpp := NIL;
         End;
       Except
 
@@ -149,35 +227,45 @@ var
   fbOK: Boolean;
   fbRetry: Boolean;
   fInternalProcRec: TPSInternalProcRec;
+  fFpp: TPSPreProcessor;
 begin
   Result := False;
 
   FCompiler := TPSPascalCompiler.Create; // create an instance of the compiler.
+
   FCompiler.OnUses := CustomOnUses; // assign the OnUses event.
 
   FCompiler.AllowNoBegin := true;
   FCompiler.AllowNoEnd := true;
-  // AllowNoBegin and AllowNoEnd allows it that begin and end are not required in a script.
+
+  fFpp := PreProcessor(Self, aprojecttask.TaskName, FScript);
 
   foAPI_Output.WriteLog('Compiling ... ');
 
   if not FCompiler.Compile(FScript.Text) then
     begin
+      fFpp.AdjustMessages(FCompiler);
+
       CompilerOutputMessage;
 
       foAPI_Output.projecttask.BuildStatus := TBuildStatus.bsFailed;
 
       FCompiler.Free;
 
+      fFpp := NIL;
+
       Exit;
     end;
 
+  fFpp.AdjustMessages(FCompiler);
+
   CompilerOutputMessage;
 
+  fFpp := NIL;
+
   FCompiler.GetOutput(FsData);
-  // Save the output of the compiler in the string Data.
+
   FCompiler.Free;
-  // After compiling the script, there is no need for the compiler anymore.
 
   if aCompileOnly then
   begin
@@ -273,4 +361,7 @@ begin
   end;
 end;
 
+
+
 end.
+